Move RequestMatcherFactory from Pimcore\Service to Pimcore\Http
