+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/aliases.yml
@@ -28,7 +28,7 @@ services:
 
     # this block is probably not needed as services are only used internally
     pimcore.service.context.pimcore_context_guesser: '@Pimcore\Service\Context\PimcoreContextGuesser'
-    pimcore.service.request_matcher_factory: '@Pimcore\Service\RequestMatcherFactory'
+    pimcore.service.request_matcher_factory: '@Pimcore\Http\RequestMatcherFactory'
 
     pimcore.locale: '@Pimcore\Localization\Locale'
     pimcore.locale.intl_formatter: '@Pimcore\Localization\IntlFormatter'
