+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/request_response.yml
@@ -8,7 +8,7 @@ services:
 
     Pimcore\Http\ResponseHelper: ~
 
-    Pimcore\Service\RequestMatcherFactory:
+    Pimcore\Http\RequestMatcherFactory:
         public: false
 
     Pimcore\Service\Context\PimcoreContextGuesser:
